requirements.txt
streamlit
tensorflow
opencv-python
numpy
pillow
matplotlib